% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_manifestation.R
\name{insert_manifestation}
\alias{insert_manifestation}
\title{Insert Manifestation Header}
\usage{
insert_manifestation(filename = NULL)
}
\arguments{
\item{filename}{Optional script name (default uses current file)}
}
\description{
Insert Manifestation Header
}
